WIP refacto common code
