+++ lib/Alchemy/Phrasea/Controller/Api/V3/V3StoriesController.php
@@ -2,7 +2,6 @@
 
 namespace Alchemy\Phrasea\Controller\Api\V3;
 
-use ACL;
 use Alchemy\Phrasea\Application\Helper\DispatcherAware;
 use Alchemy\Phrasea\Application\Helper\JsonBodyAware;
 use Alchemy\Phrasea\Controller\Api\Result;
@@ -10,12 +9,9 @@ use Alchemy\Phrasea\Controller\Api\V1Controller;
 use Alchemy\Phrasea\Controller\Controller;
 use Alchemy\Phrasea\Model\RecordReferenceInterface;
 use Alchemy\Phrasea\Record\RecordReferenceCollection;
-use caption_field;
 use caption_record;
 use databox_Field_DCESAbstract;
 use Exception;
-use media_Permalink_Adapter;
-use media_subdef;
 use record_adapter;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -49,23 +45,6 @@ class V3StoriesController extends Controller
         catch (Exception $e) {
             return $this->app['controller.api.v1']->getBadRequestAction($request, 'An error occurred');
         }
-    }
-
-    /**
-     * @return V3ResultHelpers
-     */
-    private function getResultHelpers()
-    {
-        static $rh = null;
-
-        if(is_null($rh)) {
-            $rh = new V3ResultHelpers(
-                $this,
-                $this->getConf(),
-                $this->app['media_accessor.subdef_url_generator']
-            );
-        }
-        return $rh;
     }
 
     /**
@@ -107,7 +86,7 @@ class V3StoriesController extends Controller
             'created_on'    => $story->getCreated()->format(DATE_ATOM),
             'collection_id' => $story->getCollectionId(),
             'base_id'       => $story->getBaseId(),
-            'thumbnail'     => $this->listEmbeddableMedia($request, $story, $story->get_thumbnail()),
+            'thumbnail'     => $this->getResultHelpers()->listEmbeddableMedia($request, $story, $story->get_thumbnail(), $this->getAclForUser()),
             'uuid'          => $story->getUuid(),
             'metadatas'     => [
                 '@entity@'       => V1Controller::OBJECT_TYPE_STORY_METADATA_BAG,
@@ -149,211 +128,19 @@ class V3StoriesController extends Controller
         foreach ($records->toRecords($this->getApplicationBox()) as $index => $record) {
             $record->setTechnicalDataSet($technicalData[$index]);
 
-            $data[$index] = $this->listRecord($request, $record);
-        }
-
-        return $data;
-    }
-    public function listEmbeddableMedia(Request $request, record_adapter $record, media_subdef $media)
-    {
-        if (!$media->is_physically_present()) {
-            return null;
-        }
-
-        if ($this->getAuthenticator()->isAuthenticated()) {
-            $acl = $this->getAclForUser();
-            if ($media->get_name() !== 'document'
-                && false === $acl->has_access_to_subdef($record, $media->get_name())
-            ) {
-                return null;
-            }
-            if ($media->get_name() === 'document'
-                && !$acl->has_right_on_base($record->getBaseId(), ACL::CANDWNLDHD)
-                && !$acl->has_hd_grant($record)
-            ) {
-                return null;
-            }
-        }
-
-        if ($media->get_permalink() instanceof media_Permalink_Adapter) {
-            $permalink = $this->listPermalink($media->get_permalink());
-        } else {
-            $permalink = null;
-        }
-
-        $urlTTL = (int) $request->get(
-            'subdef_url_ttl',
-            $this->getConf()->get(['registry', 'general', 'default-subdef-url-ttl'])
-        );
-        if ($urlTTL < 0) {
-            $urlTTL = -1;
-        }
-        $issuer = $this->getAuthenticatedUser();
-
-        return [
-            'name' => $media->get_name(),
-            'permalink' => $permalink,
-            'height' => $media->get_height(),
-            'width' => $media->get_width(),
-            'filesize' => $media->get_size(),
-            'devices' => $media->getDevices(),
-            'player_type' => $media->get_type(),
-            'mime_type' => $media->get_mime(),
-            'substituted' => $media->is_substituted(),
-            'created_on'  => $media->get_creation_date()->format(DATE_ATOM),
-            'updated_on'  => $media->get_modification_date()->format(DATE_ATOM),
-            'url' => $this->app['media_accessor.subdef_url_generator']->generate($issuer, $media, $urlTTL),
-            'url_ttl' => $urlTTL,
-        ];
-    }
-
-    private function listPermalink(media_Permalink_Adapter $permalink)
-    {
-        $downloadUrl = $permalink->get_url();
-        $downloadUrl->getQuery()->set('download', '1');
-
-        return [
-            'created_on'   => $permalink->get_created_on()->format(DATE_ATOM),
-            'id'           => $permalink->get_id(),
-            'is_activated' => $permalink->get_is_activated(),
-            /** @Ignore */
-            'label'        => $permalink->get_label(),
-            'updated_on'   => $permalink->get_last_modified()->format(DATE_ATOM),
-            'page_url'     => $permalink->get_page(),
-            'download_url' => (string)$downloadUrl,
-            'url'          => (string)$permalink->get_url(),
-        ];
-    }
-
-
-    /**
-     * Retrieve detailed information about one record
-     *
-     * @param Request          $request
-     * @param record_adapter $record
-     * @return array
-     */
-    public function listRecord(Request $request, record_adapter $record)
-    {
-        $technicalInformation = [];
-        foreach ($record->get_technical_infos()->getValues() as $name => $value) {
-            $technicalInformation[] = ['name' => $name, 'value' => $value];
-        }
-
-        $data = [
-            'databox_id'             => $record->getDataboxId(),
-            'record_id'              => $record->getRecordId(),
-            'mime_type'              => $record->getMimeType(),
-            'title'                  => $record->get_title(),
-            'original_name'          => $record->get_original_name(),
-            'updated_on'             => $record->getUpdated()->format(DATE_ATOM),
-            'created_on'             => $record->getCreated()->format(DATE_ATOM),
-            'collection_id'          => $record->getCollectionId(),
-            'base_id'                => $record->getBaseId(),
-            'sha256'                 => $record->getSha256(),
-            'thumbnail'              => $this->listEmbeddableMedia($request, $record, $record->get_thumbnail()),
-            'technical_informations' => $technicalInformation,
-            'phrasea_type'           => $record->getType(),
-            'uuid'                   => $record->getUuid(),
-        ];
-
-        if ($request->attributes->get('_extended', false)) {
-            $data = array_merge($data, [
-                'subdefs' => $this->listRecordEmbeddableMedias($request, $record),
-                'metadata' => $this->listRecordMetadata($record),
-                'status' => $this->getResultHelpers()->listRecordStatus($record),
-                'caption' => $this->listRecordCaption($record),
-            ]);
+            $data[$index] = $this->getResultHelpers()->listRecord($request, $record, $this->getAclForUser());
         }
 
         return $data;
     }
 
-    /**
-     * @param Request $request
-     * @param record_adapter $record
-     * @return array
-     */
-    private function listRecordEmbeddableMedias(Request $request, record_adapter $record)
-    {
-        $subdefs = [];
-
-        foreach ($record->get_embedable_medias([], []) as $name => $media) {
-            if (null !== $subdef = $this->listEmbeddableMedia($request, $record, $media)) {
-                $subdefs[] = $subdef;
-            }
-        }
-
-        return $subdefs;
-    }
 
     /**
-     * List all fields of given record
-     *
-     * @param record_adapter $record
-     * @return array
+     * @return V3ResultHelpers
      */
-    private function listRecordMetadata(record_adapter $record)
+    private function getResultHelpers()
     {
-        $includeBusiness = $this->getAclForUser()->can_see_business_fields($record->getDatabox());
-
-        return $this->listRecordCaptionFields($record->get_caption()->get_fields(null, $includeBusiness));
-    }
-
-    /**
-     * @param caption_field[] $fields
-     * @return array
-     */
-    private function listRecordCaptionFields($fields)
-    {
-        $ret = [];
-
-        foreach ($fields as $field) {
-            $databox_field = $field->get_databox_field();
-
-            $fieldData = [
-                'meta_structure_id' => $field->get_meta_struct_id(),
-                'name' => $field->get_name(),
-                'labels' => [
-                    'fr' => $databox_field->get_label('fr'),
-                    'en' => $databox_field->get_label('en'),
-                    'de' => $databox_field->get_label('de'),
-                    'nl' => $databox_field->get_label('nl'),
-                ],
-            ];
-
-            foreach ($field->get_values() as $value) {
-                $data = [
-                    'meta_id' => $value->getId(),
-                    'value' => $value->getValue(),
-                ];
-
-                $ret[] = $fieldData + $data;
-            }
-        }
-
-        return $ret;
-    }
-
-    /**
-     * @param record_adapter $record
-     * @return array
-     */
-    private function listRecordCaption(record_adapter $record)
-    {
-        $includeBusiness = $this->getAclForUser()->can_see_business_fields($record->getDatabox());
-
-        $caption = [];
-
-        foreach ($record->get_caption()->get_fields(null, $includeBusiness) as $field) {
-            $caption[] = [
-                'meta_structure_id' => $field->get_meta_struct_id(),
-                'name' => $field->get_name(),
-                'value' => $field->get_serialized_values(';'),
-            ];
-        }
-
-        return $caption;
+        return $this->app['controller.api.v3.resulthelpers'];
     }
 
 }
